+++ lib/network.dart
@@ -3,7 +3,8 @@ library illuminate;
 export 'network/src/utils.dart' show storageKeyAccessToken, storageKeyRefreshToken;
 
 export 'network/src/pkce.dart' show PKCEPair;
-export 'network/src/client.dart' show Client;
-export 'network/src/exceptions/exceptions.dart' show OAuthConfigurationException, OAuthSignInCanceledException, OAuthSignInFailureException, ResponseException;
+export 'network/src/client.dart' show Client, AuthenticationDelegate, OAuthAuthenticationDelegate;
+export 'network/src/exceptions/exceptions.dart'
+    show OAuthConfigurationException, OAuthSignInCanceledException, OAuthSignInFailureException, ResponseException;
 export 'network/src/oauth_authenticator.dart' show OAuthAuthenticator;
 export 'network/src/types.dart' show HttpMethod, OAuthConfig, ApiConfig, Request, AuthorizationRequest, Authentication;